Fixed reported errors

+++ src/Appwrite/Storage/Storage.php
@@ -104,7 +104,7 @@ class Storage
             ),
         );
 
-        $factor = floor((strlen($bytes) - 1) / 3);
+        $factor = (int)floor((strlen($bytes) - 1) / 3);
 
         return sprintf("%.{$decimals}f%s", $bytes / pow($mod, $factor), $units[$system][$factor]);
     }
